+++ src/nextjs/templates.ts
@@ -36,6 +36,12 @@ export function getNextjsSentryBuildOptionsTemplate(): string {
 
     // Automatically tree-shake Sentry logger statements to reduce bundle size
     disableLogger: true,
+
+    // Enables automatic instrumentation of Vercel Cron Monitors.
+    // See the following for more information:
+    // https://docs.sentry.io/product/crons/
+    // https://vercel.com/docs/cron-jobs
+    automaticVercelMonitors: true,
   }`;
 }
 
@@ -323,4 +329,75 @@ YourCustomErrorComponent.getInitialProps = async (contextData${
   await Sentry.captureUnderscoreErrorException(contextData);
 };
 `;
+}
+
+export function getSentryDefaultGlobalErrorPage() {
+  return `"use client";
+
+import * as Sentry from "@sentry/nextjs";
+import Error from "next/error";
+import { useEffect } from "react";
+
+export default function GlobalError({ error }) {
+  useEffect(() => {
+    Sentry.captureException(error);
+  }, [error]);
+
+  return (
+    <html>
+      <body>
+        <Error />
+      </body>
+    </html>
+  );
+}
+`;
+}
+
+export function getGlobalErrorCopyPasteSnippet(isTs: boolean) {
+  if (isTs) {
+    return `"use client";
+
+${chalk.green('import * as Sentry from "@sentry/nextjs";')}
+${chalk.green('import Error from "next/error";')}
+${chalk.green('import { useEffect } from "react";')}
+
+export default function GlobalError(${chalk.green(
+      '{ error }: { error: Error }',
+    )}) {
+  ${chalk.green(`useEffect(() => {
+    Sentry.captureException(error);
+  }, [error]);`)}
+
+  return (
+    <html>
+      <body>
+        {/* Your Error component here... */}
+      </body>
+    </html>
+  );
+}
+`;
+  } else {
+    return `"use client";
+
+${chalk.green('import * as Sentry from "@sentry/nextjs";')}
+${chalk.green('import Error from "next/error";')}
+${chalk.green('import { useEffect } from "react";')}
+
+export default function GlobalError(${chalk.green('{ error }')}) {
+  ${chalk.green(`useEffect(() => {
+    Sentry.captureException(error);
+  }, [error]);`)}
+
+  return (
+    <html>
+      <body>
+        {/* Your Error component here... */}
+      </body>
+    </html>
+  );
+}
+`;
+  }
 }